refactor: make it possible to use the encoding config without the app
